backend/requirements.txt
coreapi
coreschema
Django
django-cors-headers
djangorestframework
pytz
urllib3
django-filter